Add ManagementView
