+++ src/app/(protected)/promotions/management/PromotionCard.tsx
