--- src/app/(protected)/promotions/new/page.tsx
@@ -1,13 +1,0 @@
-// src/app/promotions/new/page.tsx
-"use client";
-
-export default function NewPromotion() {
-
-
-
-    return (
-        <div className="max-w-6xl mx-auto flex gap-6 p-6">
-            dfd
-        </div>
-    );
-}
